Create README.md
Extracted data from Twitter (unstructured data) for types of breeds of dogs:
1)Golden Retriever  2) Labrador retriever  3)Husky  4)Corgi  5)Beagle 
and visualizing the type of emojis used for them and ranking them in order. I am also trying compare them in couple and based on my analysis, Golden retriever is the most famous dogs with the highest number of smileys used. And co-incidentally, my most favorite breed is Golden Retriever too..